Split up checks.c to hosts and service check modules
Hosts and services behaves quite similar, and should be treated
equally, and by the same methods. But they don't, so don't pretend to
be the same module.

Splitting up hosts checking and service checking in different files
makes it much easier to graps one module at the same time, so we can
refactor the scheduling of one module at the time to.
