+++ tests/Makefile.am
@@ -7,18 +7,22 @@ TESTS = test-checks test-utils test-log test-config
 TEST_CHECKS_DEPS = nebmods.o commands.o broker.o query-handler.o utils.o events.o notifications.o \
 			  flapping.o sehandlers.o logging.o workers.o shared.o comments.o downtime.o sretention.o objects.o \
 			  macros.o statusdata.o xrddefault.o xsddefault.o xpddefault.o perfdata.o xodtemplate.o nm_alloc.o
-test_checks_SOURCES	= test-checks.c $(top_srcdir)/naemon/checks.h $(top_srcdir)/naemon/checks.c $(top_srcdir)/naemon/defaults.c
+test_checks_SOURCES	= test-checks.c \
+	 $(top_srcdir)/naemon/checks.h $(top_srcdir)/naemon/checks.c \
+	 $(top_srcdir)/naemon/checks_host.h $(top_srcdir)/naemon/checks_host.c \
+	 $(top_srcdir)/naemon/checks_service.h $(top_srcdir)/naemon/checks_service.c \
+	 $(top_srcdir)/naemon/defaults.c
 test_checks_LDADD =  $(TEST_CHECKS_DEPS:%=$(top_builddir)/naemon/naemon-%) $(LDADD)
 
-TEST_UTILS_DEPS = $(TEST_CHECKS_DEPS) checks.o
+TEST_UTILS_DEPS = $(TEST_CHECKS_DEPS) checks.o checks_host.o checks_service.o
 test_utils_SOURCES	= test-utils.c $(top_srcdir)/naemon/defaults.c
 test_utils_LDADD =  $(TEST_UTILS_DEPS:%=$(top_builddir)/naemon/naemon-%) $(LDADD)
 
-TEST_LOG_DEPS = $(TEST_CHECKS_DEPS) checks.o
+TEST_LOG_DEPS = $(TEST_CHECKS_DEPS) checks.o checks_host.o checks_service.o
 test_log_SOURCES	= test-log.c $(top_srcdir)/naemon/defaults.c
 test_log_LDADD = $(TEST_LOG_DEPS:%=$(top_builddir)/naemon/naemon-%) $(LDADD)
 
-TEST_CONFIG_DEPS = $(TEST_CHECKS_DEPS) checks.o configuration.o
+TEST_CONFIG_DEPS = $(TEST_CHECKS_DEPS) checks.o checks_host.o checks_service.o configuration.o
 test_config_SOURCES	= test-config.c $(top_srcdir)/naemon/defaults.c
 test_config_LDADD = $(TEST_CONFIG_DEPS:%=$(top_builddir)/naemon/naemon-%) $(LDADD)
